fixes line item delete issue

+++ common/models/InvoiceLineItem.php
@@ -284,13 +284,16 @@ class InvoiceLineItem extends \yii\db\ActiveRecord
     public function beforeSoftDelete()
     {
         if ($this->invoice->isInvoice() && $this->isLessonItem()) {
-            $lessonCreditPayments = $this->lesson->getCreditUsedPayment($this->enrolment->id);
-            foreach ($lessonCreditPayments as $lessonCreditPayment) {
-                $lessonCreditPayment->delete();
+            if ($this->lesson->hasCreditUsed($this->enrolment->id)) {
+                $lessonCreditPayments = $this->lesson->getCreditUsedPayment($this->enrolment->id);
+                foreach ($lessonCreditPayments as $lessonCreditPayment) {
+                    $lessonCreditPayment->delete();
+                }
             }
+            $this->lesson->unschedule();
         }
 
-        return $this->lesson->unschedule();
+        return true;
     }
 
     public function isDefaultDiscountedItems()
